Move new types to consts subdirectory, fix casing for various constants
