+++ src/app/consts/kind.rs
@@ -4,7 +4,7 @@ pub enum FighterKind {
     Donkey = 0x1,
     Link = 0x2,
     Samus = 0x3,
-    Samusd = 0x4,
+    SamusD = 0x4,
     Yoshi = 0x5,
     Kirby = 0x6,
     Fox = 0x7,
@@ -18,7 +18,7 @@ pub enum FighterKind {
     Koopa = 0xF,
     Sheik = 0x10,
     Zelda = 0x11,
-    Mariod = 0x12,
+    MarioD = 0x12,
     Pichu = 0x13,
     Falco = 0x14,
     Marth = 0x15,
@@ -28,17 +28,17 @@ pub enum FighterKind {
     Mewtwo = 0x19,
     Roy = 0x1A,
     Chrom = 0x1B,
-    Gamewatch = 0x1C,
+    GameWatch = 0x1C,
     Metaknight = 0x1D,
     Pit = 0x1E,
-    Pitb = 0x1F,
-    Szerosuit = 0x20,
+    PitB = 0x1F,
+    SZerosuit = 0x20,
     Wario = 0x21,
     Snake = 0x22,
     Ike = 0x23,
-    Pzenigame = 0x24,
-    Pfushigisou = 0x25,
-    Plizardon = 0x26,
+    PZenigame = 0x24,
+    PFushigisou = 0x25,
+    PLizardon = 0x26,
     Diddy = 0x27,
     Lucas = 0x28,
     Sonic = 0x29,
@@ -58,7 +58,7 @@ pub enum FighterKind {
     Pacman = 0x37,
     Reflet = 0x38,
     Shulk = 0x39,
-    Koopajr = 0x3A,
+    KoopaJr = 0x3A,
     Duckhunt = 0x3B,
     Ryu = 0x3C,
     Ken = 0x3D,
@@ -72,15 +72,15 @@ pub enum FighterKind {
     Krool = 0x45,
     Shizue = 0x46,
     Gaogaen = 0x47,
-    Miifighter = 0x48,
-    Miiswordsman = 0x49,
-    Miigunner = 0x4A,
+    MiiFighter = 0x48,
+    MiiSwordsman = 0x49,
+    MiiGunner = 0x4A,
     Popo = 0x4B,
     Nana = 0x4C,
-    Koopag = 0x4D,
-    Miienemyf = 0x4E,
-    Miienemys = 0x4F,
-    Miienemyg = 0x50,
+    KoopaG = 0x4D,
+    MiiEnemyF = 0x4E,
+    MiiEnemyS = 0x4F,
+    MiiEnemyG = 0x50,
     Packun = 0x51,
     Jack = 0x52,
     Brave = 0x53,
@@ -90,8 +90,8 @@ pub enum FighterKind {
     Tantan = 0x57,
     Pickel = 0x58,
     Edge = 0x59,
-    Eflame = 0x5A,
-    Elight = 0x5B,
+    EFlame = 0x5A,
+    ELight = 0x5B,
     Demon = 0x5C,
     Trail = 0x5D,
     TableExStart = 0x6D,
@@ -99,7 +99,7 @@ pub enum FighterKind {
     Zenigame = 0x6F,
     Fushigisou = 0x70,
     Lizardon = 0x71,
-    Ptrainer = 0x72,
+    PTrainer = 0x72,
     Flame = 0x73,
     Light = 0x74,
     Element = 0x75,
@@ -113,8 +113,8 @@ impl FighterKind {
     pub const BASE_TAIL: FighterKind = FighterKind::Nana;
     pub const BASE_NUM: usize        = 0x4D;
     
-    pub const OTHER_HEAD: FighterKind = FighterKind::Koopag;
-    pub const OTHER_TAIL: FighterKind = FighterKind::Miienemyg;
+    pub const OTHER_HEAD: FighterKind = FighterKind::KoopaG;
+    pub const OTHER_TAIL: FighterKind = FighterKind::MiiEnemyG;
     pub const OTHER_NUM: usize        = 0x4;
 
     pub const BASE_APPEND_HEAD: FighterKind = FighterKind::Packun;
@@ -125,27 +125,27 @@ impl FighterKind {
 #[repr(u32)]
 pub enum WeaponKind {
     None = 0xFFFFFFFF,
-    MiifighterHat = 0x0,
-    MiifighterIronball = 0x1,
-    MiiswordsmanHat = 0x2,
-    MiiswordsmanLightshuriken = 0x3,
-    MiiswordsmanChakram = 0x4,
-    MiiswordsmanTornadoshot = 0x5,
-    MiiswordsmanWave = 0x6,
-    MiigunnerHat = 0x7,
-    MiigunnerRapidshotBullet = 0x8,
-    MiigunnerGunnercharge = 0x9,
-    MiigunnerMiimissile = 0xA,
-    MiigunnerSupermissile = 0xB,
-    MiigunnerGrenadelauncher = 0xC,
-    MiigunnerFlamepillar = 0xD,
-    MiigunnerStealthbomb = 0xE,
-    MiigunnerStealthbombS = 0xF,
-    MiigunnerBottomshoot = 0x10,
-    MiigunnerGroundbomb = 0x11,
-    MiigunnerAttackairfBullet = 0x12,
-    MiigunnerLaser = 0x13,
-    MiigunnerFullthrottle = 0x14,
+    MiiFighterHat = 0x0,
+    MiiFighterIronball = 0x1,
+    MiiSwordsmanHat = 0x2,
+    MiiSwordsmanLightshuriken = 0x3,
+    MiiSwordsmanChakram = 0x4,
+    MiiSwordsmanTornadoshot = 0x5,
+    MiiSwordsmanWave = 0x6,
+    MiiGunnerHat = 0x7,
+    MiiGunnerRapidshotBullet = 0x8,
+    MiiGunnerGunnercharge = 0x9,
+    MiiGunnerMiimissile = 0xA,
+    MiiGunnerSupermissile = 0xB,
+    MiiGunnerGrenadelauncher = 0xC,
+    MiiGunnerFlamepillar = 0xD,
+    MiiGunnerStealthbomb = 0xE,
+    MiiGunnerStealthbombS = 0xF,
+    MiiGunnerBottomshoot = 0x10,
+    MiiGunnerGroundbomb = 0x11,
+    MiiGunnerAttackairfBullet = 0x12,
+    MiiGunnerLaser = 0x13,
+    MiiGunnerFullthrottle = 0x14,
     MarioFireball = 0x15,
     MarioMantle = 0x16,
     MarioPump = 0x17,
@@ -171,16 +171,16 @@ pub enum WeaponKind {
     SamusSupermissile = 0x2B,
     SamusTransportation = 0x2C,
     SamusGbeam = 0x2D,
-    SamusdCshot = 0x2E,
-    SamusdBomb = 0x2F,
-    SamusdMissile = 0x30,
-    SamusdLaser = 0x31,
-    SamusdLaser2 = 0x32,
-    SamusdGun = 0x33,
-    SamusdSupermissile = 0x34,
-    SamusdTransportation = 0x35,
-    SamusdGbeam = 0x36,
-    SamusdBunshin = 0x37,
+    SamusDCshot = 0x2E,
+    SamusDBomb = 0x2F,
+    SamusDMissile = 0x30,
+    SamusDLaser = 0x31,
+    SamusDLaser2 = 0x32,
+    SamusDGun = 0x33,
+    SamusDSupermissile = 0x34,
+    SamusDTransportation = 0x35,
+    SamusDGbeam = 0x36,
+    SamusDBunshin = 0x37,
     YoshiStar = 0x38,
     YoshiTamago = 0x39,
     YoshiYoshimob = 0x3A,
@@ -235,7 +235,7 @@ pub enum WeaponKind {
     DaisyKinopio = 0x6B,
     DaisyKinopiospore = 0x6C,
     KoopaBreath = 0x6D,
-    KoopaKoopag = 0x6E,
+    KoopaKoopaG = 0x6E,
     ZeldaDein = 0x6F,
     ZeldaDeinS = 0x70,
     ZeldaPhantom = 0x71,
@@ -244,19 +244,19 @@ pub enum WeaponKind {
     SheikNeedlehave = 0x74,
     SheikFusin = 0x75,
     SheikKnife = 0x76,
-    GamewatchFood = 0x77,
-    GamewatchRescue = 0x78,
-    GamewatchParachute = 0x79,
-    GamewatchPanel = 0x7A,
-    GamewatchOil = 0x7B,
-    GamewatchOctopus = 0x7C,
-    GamewatchNormalWeapon = 0x7D,
-    GamewatchBreath = 0x7E,
-    GamewatchEntry = 0x7F,
-    GamewatchBomb = 0x80,
+    GameWatchFood = 0x77,
+    GameWatchRescue = 0x78,
+    GameWatchParachute = 0x79,
+    GameWatchPanel = 0x7A,
+    GameWatchOil = 0x7B,
+    GameWatchOctopus = 0x7C,
+    GameWatchNormalWeapon = 0x7D,
+    GameWatchBreath = 0x7E,
+    GameWatchEntry = 0x7F,
+    GameWatchBomb = 0x80,
     GanonSword = 0x81,
     GanonBeast = 0x82,
-    GanonGanond = 0x83,
+    GanonGanonD = 0x83,
     FalcoBlaster = 0x84,
     FalcoBlasterBullet = 0x85,
     FalcoIllusion = 0x86,
@@ -274,14 +274,14 @@ pub enum WeaponKind {
     PitChariot = 0x92,
     PitHorse = 0x93,
     PitChariotsight = 0x94,
-    SzerosuitParalyzerBullet = 0x95,
-    SzerosuitWhip = 0x96,
-    SzerosuitParalyzer = 0x97,
-    SzerosuitSamusp = 0x98,
-    SzerosuitGunship = 0x99,
-    SzerosuitReticle = 0x9A,
-    SzerosuitLaser = 0x9B,
-    SzerosuitWhip2 = 0x9C,
+    SZerosuitParalyzerBullet = 0x95,
+    SZerosuitWhip = 0x96,
+    SZerosuitParalyzer = 0x97,
+    SZerosuitSamusp = 0x98,
+    SZerosuitGunship = 0x99,
+    SZerosuitReticle = 0x9A,
+    SZerosuitLaser = 0x9B,
+    SZerosuitWhip2 = 0x9C,
     PikminPikmin = 0x9D,
     PikminDolfin = 0x9E,
     PikminWin1 = 0x9F,
@@ -307,7 +307,7 @@ pub enum WeaponKind {
     IkeSword = 0xB3,
     LucarioAuraball = 0xB4,
     LucarioQigong = 0xB5,
-    LucarioLucariom = 0xB6,
+    LucarioLucarioM = 0xB6,
     RobotGyro = 0xB7,
     RobotGyroHolder = 0xB8,
     RobotBeam = 0xB9,
@@ -331,20 +331,20 @@ pub enum WeaponKind {
     WolfIllusion = 0xCB,
     WolfWolfen = 0xCC,
     WolfReticle = 0xCD,
-    PlizardonBreath = 0xCE,
-    PlizardonExplosion = 0xCF,
-    PlizardonDaimonji = 0xD0,
+    PLizardonBreath = 0xCE,
+    PLizardonExplosion = 0xCF,
+    PLizardonDaimonji = 0xD0,
     SonicHomingtarget = 0xD1,
     SonicGimmickjump = 0xD2,
     SonicSupersonic = 0xD3,
     SonicChaosemerald = 0xD4,
     PurinCap = 0xD5,
     PurinMonsterball = 0xD6,
-    MariodDrcapsule = 0xD7,
-    MariodDrmantle = 0xD8,
-    MariodStethoscope = 0xD9,
-    MariodHugeCapsule = 0xDA,
-    MariodCapsuleblock = 0xDB,
+    MarioDDrcapsule = 0xD7,
+    MarioDDrmantle = 0xD8,
+    MarioDStethoscope = 0xD9,
+    MarioDHugeCapsule = 0xDA,
+    MarioDCapsuleblock = 0xDB,
     PichuDengekidama = 0xDC,
     PichuDengeki = 0xDD,
     PichuKaminari = 0xDE,
@@ -353,8 +353,8 @@ pub enum WeaponKind {
     PichuMonsterball = 0xE1,
     PichuSpecialupdummy = 0xE2,
     LucinaMask = 0xE3,
-    PitbBow = 0xE4,
-    PitbBowarrow = 0xE5,
+    PitBBow = 0xE4,
+    PitBBowarrow = 0xE5,
     RosettaTico = 0xE6,
     RosettaStarpiece = 0xE7,
     RosettaRing = 0xE8,
@@ -367,12 +367,12 @@ pub enum WeaponKind {
     WiifitWiibo = 0xEF,
     WiifitTowel = 0xF0,
     WiifitSilhouette = 0xF1,
-    WiifitSilhouettel = 0xF2,
+    WiifitSilhouetteL = 0xF2,
     LittlemacDoclouis = 0xF3,
     LittlemacSweatlittlemac = 0xF4,
     LittlemacThrowsweat = 0xF5,
     LittlemacChampionbelt = 0xF6,
-    LittlemacLittlemacg = 0xF7,
+    LittlemacLittlemacG = 0xF7,
     MurabitoFlowerpot = 0xF8,
     MurabitoUmbrella = 0xF9,
     MurabitoBowlingBall = 0xFA,
@@ -425,14 +425,14 @@ pub enum WeaponKind {
     DuckhuntFinalgrass = 0x129,
     DuckhuntFinalcan = 0x12A,
     DuckhuntGrass = 0x12B,
-    KoopajrHammer = 0x12C,
-    KoopajrPicopicohammer = 0x12D,
-    KoopajrMagichand = 0x12E,
-    KoopajrCannonball = 0x12F,
-    KoopajrKart = 0x130,
-    KoopajrRemainclown = 0x131,
-    KoopajrShadowmario = 0x132,
-    KoopajrBatten = 0x133,
+    KoopaJrHammer = 0x12C,
+    KoopaJrPicopicohammer = 0x12D,
+    KoopaJrMagichand = 0x12E,
+    KoopaJrCannonball = 0x12F,
+    KoopaJrKart = 0x130,
+    KoopaJrRemainclown = 0x131,
+    KoopaJrShadowmario = 0x132,
+    KoopaJrBatten = 0x133,
     ShulkDunban = 0x134,
     ShulkRiki = 0x135,
     ShulkFiora = 0x136,
@@ -442,7 +442,7 @@ pub enum WeaponKind {
     GekkougaMoon = 0x13A,
     GekkougaMonsterball = 0x13B,
     GekkougaBunshin = 0x13C,
-    GekkougaGekkougas = 0x13D,
+    GekkougaGekkougaS = 0x13D,
     PacmanEsa = 0x13E,
     PacmanTrampoline = 0x13F,
     PacmanFirehydrant = 0x140,
@@ -468,7 +468,7 @@ pub enum WeaponKind {
     RockmanAirshooter = 0x154,
     MewtwoShadowball = 0x155,
     MewtwoBindball = 0x156,
-    MewtwoMewtwom = 0x157,
+    MewtwoMewtwoM = 0x157,
     MewtwoSearch = 0x158,
     MewtwoPsychobreak = 0x159,
     MewtwoEscapeairdummy = 0x15A,
@@ -510,10 +510,10 @@ pub enum WeaponKind {
     PopoIcebergWind = 0x17E,
     PopoIcebergHit = 0x17F,
     PopoWhitebear = 0x180,
-    PzenigameWater = 0x181,
-    PfushigisouSeed = 0x182,
-    PfushigisouLeafcutter = 0x183,
-    PfushigisouVine = 0x184,
+    PZenigameWater = 0x181,
+    PFushigisouSeed = 0x182,
+    PFushigisouLeafcutter = 0x183,
+    PFushigisouVine = 0x184,
     InklingInkbullet = 0x185,
     InklingSplashbomb = 0x186,
     InklingSquid = 0x187,
@@ -566,7 +566,7 @@ pub enum WeaponKind {
     ChromSword = 0x1B6,
     RidleyBreath = 0x1B7,
     RidleyGunship = 0x1B8,
-    KoopagBreath = 0x1B9,
+    KoopaGBreath = 0x1B9,
     GaogaenRope = 0x1BA,
     GaogaenChampionbelt = 0x1BB,
     GaogaenMonsterball = 0x1BC,
@@ -581,13 +581,13 @@ pub enum WeaponKind {
     RichterWhipwire = 0x1C5,
     RichterStake = 0x1C6,
     EnemyYellowdevilBeam = 0x1C7,
-    MiienemygAttackairfBullet = 0x1C8,
-    MiienemygRapidshotBullet = 0x1C9,
-    PtrainerPtrainer = 0x1CA,
-    PtrainerMball = 0x1CB,
-    PtrainerPzenigame = 0x1CC,
-    PtrainerPfushigisou = 0x1CD,
-    PtrainerPlizardon = 0x1CE,
+    MiiEnemyGAttackairfBullet = 0x1C8,
+    MiiEnemyGRapidshotBullet = 0x1C9,
+    PTrainerPTrainer = 0x1CA,
+    PTrainerMball = 0x1CB,
+    PTrainerPZenigame = 0x1CC,
+    PTrainerPFushigisou = 0x1CD,
+    PTrainerPLizardon = 0x1CE,
     ElementChanger = 0x1CF,
     ElementDiver = 0x1D0,
     ShizuePot = 0x1D1,
@@ -713,21 +713,21 @@ pub enum WeaponKind {
     EdgeFlare2 = 0x249,
     EdgeFlaredummy = 0x24A,
     EdgeBackground = 0x24B,
-    EflameEsword = 0x24C,
-    EflameFirepillar = 0x24D,
-    EflameBlazepillar = 0x24E,
-    EflameWindummy = 0x24F,
-    EflameWinsword = 0x250,
-    ElightEsword = 0x251,
-    ElightBunshin = 0x252,
-    ElightExprosiveshot = 0x253,
-    ElightSpreadbullet = 0x254,
-    ElightMeteor = 0x255,
-    ElightBeam = 0x256,
-    ElightWindummy = 0x257,
-    ElightWinsword = 0x258,
+    EFlameEsword = 0x24C,
+    EFlameFirepillar = 0x24D,
+    EFlameBlazepillar = 0x24E,
+    EFlameWindummy = 0x24F,
+    EFlameWinsword = 0x250,
+    ELightEsword = 0x251,
+    ELightBunshin = 0x252,
+    ELightExprosiveshot = 0x253,
+    ELightSpreadbullet = 0x254,
+    ELightMeteor = 0x255,
+    ELightBeam = 0x256,
+    ELightWindummy = 0x257,
+    ELightWinsword = 0x258,
     DemonBlaster = 0x259,
-    DemonDemonp = 0x25A,
+    DemonDemonP = 0x25A,
     DemonBlasterchest = 0x25B,
     DemonBlasterhead = 0x25C,
     DemonBlasterwing = 0x25D,
@@ -1143,7 +1143,7 @@ pub enum ItemKind {
     Lioleusbossbarrel = 0x18B,
     Lioleusbossbombhei = 0x18C,
     Lioleusbosssmartbomb = 0x18D,
-    Koopag = 0x18E,
+    KoopaG = 0x18E,
     Kiila = 0x18F,
     Kiilacrossbomb = 0x190,
     Kiilacrushwing = 0x191,
@@ -1579,7 +1579,7 @@ pub enum ItemKind {
     ForDuetLioleusbossbarrel = 0x340,
     ForDuetLioleusbossbombhei = 0x341,
     ForDuetLioleusbosssmartbomb = 0x342,
-    ForDuetKoopag = 0x343,
+    ForDuetKoopaG = 0x343,
     ForDuetKiila = 0x344,
     ForDuetKiilacrossbomb = 0x345,
     ForDuetKiilacrushwing = 0x346,
@@ -1622,4 +1622,24 @@ pub enum ItemKind {
 
 impl ItemKind {
     pub const LOT_START: ItemKind = ItemKind::ExRndAny;
+}
+
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+#[repr(u8)]
+pub enum CameraQuakeKind {
+    None = 0x0,
+    KeepSmall = 0x1,
+    SmallHalf = 0x2,
+    Small = 0x3,
+    KeepMiddle = 0x4,
+    Middle = 0x5,
+    KeepLarge = 0x6,
+    Large = 0x7,
+    KeepMoreLarge = 0x8,
+    MoreLarge = 0x9,
+    PowBlock = 0xA,
+    Knockout = 0xB,
+    DollyStage = 0xC,
+    XXL = 0xD,
+    Invalid = 0xE,
 }
